requirements.txt
pytest-playwright==0.4.2
pytest-html